pkg/apis/monitoring/v1: move each CRD types into each own file
