+++ pkg/apis/monitoring/v1/thanos_ruler_types.go
